requirements.txt
# IoT & MQTT
paho-mqtt>=2.0.0

# ML Core
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Deep Learning (optional — needed for LSTM forecaster)
tensorflow>=2.15.0
